refactor(5746): update Airflow codebase for compatibility with v3
