+++ helm/tools/dags/acs_dev.py
@@ -11,15 +11,13 @@ MONGO_CONN_ID = "pltsvc-dev"
 
 with DAG(
     dag_id="acs_dev",
-    schedule_interval="0 2 * * *",
+    schedule="0 2 * * *",
     start_date=YESTERDAY,
-    concurrency=CONCURRENCY,
+    max_active_tasks=CONCURRENCY,
 ) as dag:
     t1 = PythonOperator(
         task_id="fetch-load-acs-projects-dev",
         python_callable=fetch_load_acs_projects,
         op_kwargs={"mongo_conn_id": MONGO_CONN_ID},
-        provide_context=True,
         on_failure_callback=lambda context: send_alert(context, "acs_dev"),
-        dag=dag,
     )
