+++ helm/tools/dags/mailchimp_subscription_dev.py
@@ -1,6 +1,6 @@
 import os
 from airflow import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from datetime import datetime, timedelta
 from _update_mailchimp_list import update_mailchimp_segment
 from _task_failure_callback import send_alert
@@ -15,7 +15,7 @@ MAILCHIMP_REGISTRY_PRIVATE_TAG_ID = os.getenv("MAILCHIMP_TEST_TAG_ID")
 with DAG(
     dag_id="mailchimp_dev",
     description="A DAG to update Mailchimp segment",
-    schedule_interval="0 5 * * *",
+    schedule="0 5 * * *",
     start_date=datetime.now() - timedelta(weeks=1),
     is_paused_upon_creation=True,
     catchup=False,
@@ -30,7 +30,5 @@ with DAG(
             "tag_id": MAILCHIMP_REGISTRY_PRIVATE_TAG_ID,
             "mongo_conn_id": MONGO_CONN_ID,
         },
-        provide_context=True,
         on_failure_callback=lambda context: send_alert(context, "mailchimp_dev"),
-        dag=dag,
     )
