+++ helm/tools/dags/provisioner_prod.py
@@ -1,6 +1,6 @@
 import os
 from airflow import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from datetime import datetime, timedelta
 from _mark_provisioned import fetch_products_mark_completed
 from _task_failure_callback import send_alert
@@ -13,9 +13,7 @@ KEYCLOAK_REALM = "platform-services"
 REGISTRY_PROVISION_SA_ID = os.getenv("PROD_PROVISION_SA_ID")
 REGISTRY_PROVISION_SA_SECRET = os.getenv("PROD_PROVISION_SA_SECRET")
 
-with DAG(
-    dag_id="provisioner_prod", schedule_interval="*/7 * * * *", start_date=datetime.now() - timedelta(minutes=8)
-) as dag:
+with DAG(dag_id="provisioner_prod", schedule="*/7 * * * *", start_date=datetime.now() - timedelta(minutes=8)) as dag:
     t1 = PythonOperator(
         task_id="fetch-products-mark-completed-prod",
         python_callable=fetch_products_mark_completed,
@@ -28,7 +26,5 @@ with DAG(
             "kc_client_id": REGISTRY_PROVISION_SA_ID,
             "kc_client_secret": REGISTRY_PROVISION_SA_SECRET,
         },
-        provide_context=True,
         on_failure_callback=lambda context: send_alert(context, "provisioner_prod"),
-        dag=dag,
     )
