+++ helm/tools/values.yaml
@@ -42,36 +42,29 @@ airflow:
   nameOverride: airflow
   useStandardNaming: true
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L46
   uid:
   gid:
 
   defaultAirflowRepository: ghcr.io/bcgov/pltsvc-tools-airflow
   # See https://github.com/bcgov/platform-services-registry/pkgs/container/pltsvc-tools-airflow
-  defaultAirflowTag: "2b73bf2443ecc9af5a6700070b255870c78d6adb" # pragma: allowlist secret
+  defaultAirflowTag: "7c50ee38729c0acff0fdd20371d7617f5f827815" # pragma: allowlist secret
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L77
   airflowVersion: "3.0.1"
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L80
   images:
     useDefaultImageForMigration: false
     migrationsWaitTimeout: 120
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L257
   rbac:
     create: true
     createSCCRoleBinding: false
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L264
   executor: KubernetesExecutor
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L370
   extraEnvFrom: | # pragma: allowlist secret
     - secretRef:
         name: airflow-variables
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L379
   data:
     metadataSecretName: secdash-postgresql # pragma: allowlist secret
     # metadataConnection:
@@ -82,10 +75,8 @@ airflow:
     #   port: 5432
     #   db: airflow
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L430
   webserverSecretKeySecretName: secdash-airflow-webserver # pragma: allowlist secret
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L490
   workers:
     extraVolumes:
       - name: secdash-airflow-dags
@@ -100,7 +91,6 @@ airflow:
       - name: secdash-airflow-shared
         mountPath: /opt/airflow/shared
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L770
   scheduler:
     replicas: 1
     resources:
@@ -124,9 +114,21 @@ airflow:
       - name: secdash-airflow-shared
         mountPath: /opt/airflow/shared
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L1130
+  dagProcessor:
+    extraVolumes:
+      - name: secdash-airflow-dags
+        persistentVolumeClaim:
+          claimName: secdash-airflow-dags
+      - name: secdash-airflow-shared
+        persistentVolumeClaim:
+          claimName: secdash-airflow-shared
+    extraVolumeMounts:
+      - name: secdash-airflow-dags
+        mountPath: /opt/airflow/dags
+      - name: secdash-airflow-shared
+        mountPath: /opt/airflow/shared
+
   webserver:
-    enabled: true
     replicas: 1
     resources:
       limits:
@@ -138,19 +140,27 @@ airflow:
     waitForMigrations:
       enabled: false
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L1342
+  apiserver:
+    replicas: 1
+    resources:
+      limits:
+        cpu: 1000m
+        memory: 2Gi
+      requests:
+        cpu: 500m
+        memory: 1Gi
+    waitForMigrations:
+      enabled: true
+
   triggerer:
     enabled: false
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L1807
   statsd:
     enabled: false
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L2278
   postgresql:
     enabled: false
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L2298
   config:
     core:
       dag_concurrency: 5
@@ -174,7 +184,6 @@ airflow:
       run_as_user: "{{ .Values.uid }}"
       delete_worker_pods: "True"
 
-  # See https://github.com/apache/airflow/blob/4a4913c6f729756b528a0f0095c0273e853bc197/chart/values.yaml#L2397
   dags:
     persistence:
       enabled: false
