+++ tools/airflow/Dockerfile
@@ -1,3 +1,3 @@
 FROM apache/airflow:3.0.1
 COPY requirements.txt /
-RUN pip install --no-cache-dir "apache-airflow==${AIRFLOW_VERSION}" -r /requirements.txt
+RUN pip install --no-cache-dir -r /requirements.txt
